Added screens folder, moved homescreen to home_screen.dart
